OGD-210 - Rename itmp-individual-details-stub to individual-details-des-stub
