+++ app/uk/gov/hmrc/individualdetailsdesstub/config/microserviceGlobal.scala
@@ -1,0 +1,70 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.individualdetailsdesstub.config
+
+import com.typesafe.config.Config
+import net.ceedubs.ficus.Ficus._
+import play.api.mvc.RequestHeader
+import play.api.{Application, Configuration, Play}
+import uk.gov.hmrc.individualdetailsdesstub.domain.ErrorBadRequest
+import uk.gov.hmrc.play.audit.filters.AuditFilter
+import uk.gov.hmrc.play.auth.controllers.AuthParamsControllerConfig
+import uk.gov.hmrc.play.auth.microservice.filters.AuthorisationFilter
+import uk.gov.hmrc.play.config.{AppName, ControllerConfig, RunMode}
+import uk.gov.hmrc.play.filters.MicroserviceFilterSupport
+import uk.gov.hmrc.play.http.logging.filters.LoggingFilter
+import uk.gov.hmrc.play.microservice.bootstrap.DefaultMicroserviceGlobal
+import scala.concurrent.ExecutionContext.Implicits.global
+
+import scala.concurrent.Future
+
+object ControllerConfiguration extends ControllerConfig {
+  lazy val controllerConfigs = Play.current.configuration.underlying.as[Config]("controllers")
+}
+
+object AuthParamsControllerConfiguration extends AuthParamsControllerConfig {
+  lazy val controllerConfigs = ControllerConfiguration.controllerConfigs
+}
+
+object MicroserviceAuditFilter extends AuditFilter with AppName with MicroserviceFilterSupport {
+  override val auditConnector = MicroserviceAuditConnector
+  override def controllerNeedsAuditing(controllerName: String) = ControllerConfiguration.paramsForController(controllerName).needsAuditing
+}
+
+object MicroserviceLoggingFilter extends LoggingFilter with MicroserviceFilterSupport {
+  override def controllerNeedsLogging(controllerName: String) = ControllerConfiguration.paramsForController(controllerName).needsLogging
+}
+
+object MicroserviceAuthFilter extends AuthorisationFilter with MicroserviceFilterSupport {
+  override lazy val authParamsConfig = AuthParamsControllerConfiguration
+  override lazy val authConnector = MicroserviceAuthConnector
+  override def controllerNeedsAuth(controllerName: String): Boolean = ControllerConfiguration.paramsForController(controllerName).needsAuth
+}
+
+object MicroserviceGlobal extends DefaultMicroserviceGlobal with RunMode with MicroserviceFilterSupport {
+  override val auditConnector = MicroserviceAuditConnector
+
+  override def microserviceMetricsConfig(implicit app: Application): Option[Configuration] = app.configuration.getConfig("microservice.metrics")
+
+  override val loggingFilter = MicroserviceLoggingFilter
+
+  override val microserviceAuditFilter = MicroserviceAuditFilter
+
+  override val authFilter = Some(MicroserviceAuthFilter)
+
+  override def onBadRequest(request: RequestHeader, error: String) = Future(ErrorBadRequest(error).toHttpResponse)
+}
