+++ conf/app.routes
@@ -1,5 +1,5 @@
 # microservice specific routes
 
-GET        /pay-as-you-earn/individuals/:ninoNoSuffix        @uk.gov.hmrc.itmpindividualdetailsstub.controller.IndividualsController.fetchOrCreateIndividual(ninoNoSuffix: NinoNoSuffix)
+GET        /pay-as-you-earn/individuals/:ninoNoSuffix        @uk.gov.hmrc.individualdetailsdesstub.controller.IndividualsController.fetchOrCreateIndividual(ninoNoSuffix: NinoNoSuffix)
 
-GET        /matching/find                                    @uk.gov.hmrc.itmpindividualdetailsstub.controller.IndividualsController.findCidPerson(nino: Nino)
+GET        /matching/find                                    @uk.gov.hmrc.individualdetailsdesstub.controller.IndividualsController.findCidPerson(nino: Nino)
