+++ conf/application.conf
@@ -15,7 +15,7 @@
 # This is the main configuration file for the application.
 # ~~~~~
 
-appName=itmp-individual-details-stub
+appName=individual-details-des-stub
 
 # Play Modules
 # ~~~~
@@ -55,7 +55,7 @@ application.langs="en"
 # ~~~~~
 # Define the Global object class for this application.
 # Default to Global in the root package.
-application.global=uk.gov.hmrc.itmpindividualdetailsstub.config.MicroserviceGlobal
+application.global= uk.gov.hmrc.individualdetailsdesstub.config.MicroserviceGlobal
 
 # Router
 # ~~~~~
@@ -86,7 +86,7 @@ controllers {
       needsAuditing = false
     }
 
-    uk.gov.hmrc.itmpindividualdetailsstub.controller.IndividualsController = {
+    uk.gov.hmrc.individualdetailsdesstub.controller.IndividualsController = {
       needsAuth = false
       needsLogging = false
       needsAuditing = false
@@ -139,7 +139,7 @@ auditing {
 
 
 mongodb {
-    uri = "mongodb://localhost:27017/itmp-individual-details-stub"
+    uri = "mongodb://localhost:27017/individual-details-des-stub"
 }
